requirements.txt
# 串口数据读取器依赖包
pyserial>=3.5